requirements.txt
aiohttp
lxml
python-dotenv